+++ src/main/kotlin/cn/berberman/emerald/nms/wrapper/chat/NmsChatMessageType.kt
@@ -1,4 +1,4 @@
-package cn.berberman.emerald.nms.data.chat
+package cn.berberman.emerald.nms.wrapper.chat
 
 import cn.berberman.emerald.util.NmsUtil
