+++ src/main/kotlin/cn/berberman/emerald/nms/wrapper/player/BukkitCraftPlayer.kt
@@ -1,4 +1,4 @@
-package cn.berberman.emerald.nms.data.player
+package cn.berberman.emerald.nms.wrapper.player
 
 import cn.berberman.emerald.reflection.ReflectionWrapper
 import cn.berberman.emerald.util.NmsUtil
