+++ src/main/kotlin/cn/berberman/emerald/nms/wrapper/player/NmsEntityPlayer.kt
@@ -1,7 +1,7 @@
-package cn.berberman.emerald.nms.data.player
+package cn.berberman.emerald.nms.wrapper.player
 
-import cn.berberman.emerald.reflection.getFieldAccess
 import cn.berberman.emerald.reflection.ReflectionWrapper
+import cn.berberman.emerald.reflection.getFieldAccess
 import cn.berberman.emerald.util.NmsUtil
 
 /**
