+++ src/main/kotlin/cn/berberman/emerald/nms/wrapper/player/NmsPacketPlayOutChat.kt
@@ -1,8 +1,8 @@
-package cn.berberman.emerald.nms.data.player
+package cn.berberman.emerald.nms.wrapper.player
 
-import cn.berberman.emerald.nms.data.chat.NmsChatMessageType
-import cn.berberman.emerald.nms.data.chat.interfaces.NmsIChatBaseComponent
-import cn.berberman.emerald.nms.data.chat.interfaces.NmsPacket
+import cn.berberman.emerald.nms.wrapper.chat.NmsChatMessageType
+import cn.berberman.emerald.nms.wrapper.chat.interfaces.NmsIChatBaseComponent
+import cn.berberman.emerald.nms.wrapper.chat.interfaces.NmsPacket
 import cn.berberman.emerald.util.NmsUtil
 
 /**
