+++ src/main/kotlin/cn/berberman/emerald/util/NmsUtil.kt
@@ -1,6 +1,6 @@
 package cn.berberman.emerald.util
 
-import cn.berberman.emerald.nms.data.item.NmsItemStack
+import cn.berberman.emerald.nms.wrapper.item.NmsItemStack
 import org.bukkit.Bukkit
 import org.bukkit.inventory.ItemStack
 import org.bukkit.plugin.IllegalPluginAccessException
@@ -31,8 +31,6 @@ object NmsUtil {
 	 * Convert bukkit ItemStack to NmsItemStack
 	 * @param original bukkit ItemStack
 	 */
-//	fun asNMSCopy(original: ItemStack): Any = getCraftBukkitClass("inventory.CraftItemStack").invokeMethod(original, "asNMSCopy", original)
-//			?: throw IllegalPluginAccessException("Convert $original error.")
 	fun asNMSCopy(original: ItemStack): Any = getCraftBukkitClass("inventory.CraftItemStack")
 			.getMethod("asNMSCopy", ItemStack::class.java)?.invoke(original, original)
 			?: throw IllegalPluginAccessException("Convert $original error.")
@@ -42,9 +40,6 @@ object NmsUtil {
 	 * Convert NMS ItemStack to bukkit ItemStack
 	 * @param original NMS ItemStack
 	 */
-//	fun asBukkitCopy(original: NmsItemStack) = getCraftBukkitClass("inventory.CraftItemStack")
-//			.invokeMethod(original.instance, "asBukkitCopy", original.instance) as? ItemStack
-//			?: throw IllegalPluginAccessException("Convert $original error.")
 	fun asBukkitCopy(original: NmsItemStack) = getCraftBukkitClass("inventory.CraftItemStack")
 			.methods.firstOrNull { it.name == "asBukkitCopy" }?.invoke(original.instance, original.instance) as? ItemStack
 			?: throw IllegalPluginAccessException("Convert $original error.")
